refactor: Moved css files from style into styles and renamed imports.
