+++ client/src/App.jsx
@@ -1,5 +1,5 @@
 import FilterProvider from './context/FilterContext';
-import './style/App.css';
+import './styles/App.css';
 import CrimeFilterMultiSelect from './components/CrimeFilterMultiSelect';
 import CommunityFilterMultiSelect from './components/CommunityFilterMultiSelect';
 import DateRangeFilter from './components/DateRangeFilter';
