+++ client/src/styles/CommunityPopup.css
